requirements.txt
langchain
streamlit
python-dotenv
langchain-groq
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
wikipedia
arxiv
langchain-openai
langchainhub